requirements.txt
asgiref==3.3.4
Django==3.2.3
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
phonenumbers==8.12.23
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
